% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{TODO - Brief description}
\usage{
analysis(data, year, setup_level)
}
\arguments{
\item{data}{dataframe containing the data to analysis}

\item{year}{character value corresponding to the the year to analyse from the data}

\item{setup_level}{character value "daily" or "weekly" dependent on how the data was sampled.}
}
\value{
TODO  describe the ouput structure here
}
\description{
TODO - Longer description here (maybe one or two sentences)
}
\examples{

\donttest{
library(ButterflyLS)

# lazy load data
data(dark_green_fritillary_weekly)
# year to analyse
year <- "2018"
# weekly sampling
setup_level <- "weekly"

# run analysis
results <- analysis(dark_green_fritillary_weekly,year,setup_level)
}

}
