% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_multiyear.R
\name{analysis_multiyear}
\alias{analysis_multiyear}
\title{TODO - Brief description}
\usage{
analysis_multiyear(data, setup_level, phi_type)
}
\arguments{
\item{data}{dataframe containing the data to analysis}

\item{setup_level}{character value "doubleslope","slope", "variable" or "fixed"}
}
\value{
TODO  describe the ouput structure here
}
\description{
TODO - Longer description here (maybe one or two sentences)
}
\examples{

\donttest{
library(ButterflyLS)

# lazy load data
data(orange_tip_weekly)
# year to analyse
year <- "2018"
# weekly sampling
setup_level <- "weekly"
# phi type
phi_type = "slope"

# run analysis
results <- analysis_multiyear(orange_tip_weekly,setup_level,phi_type)
}

}
