% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_green_fritillary_daily.R
\docType{data}
\name{dark_green_fritillary_daily}
\alias{dark_green_fritillary_daily}
\title{Dark Green Fritillary}
\format{
A dataframe with 5 columns and 581129 rows. The columns are DAYNO (day number), WEEKNO (week number), SITENO (site number), YEAR (year), SPECIES (species identify), COUNT (count), and DATE (date).
}
\usage{
data(dark_green_fritillary_daily )
}
\description{
Daily site, year, and count data for dark green fritillary
}
\examples{

library(ButterflyLS)

data(dark_green_fritillary_daily )
head(dark_green_fritillary_daily )
plot(dark_green_fritillary_daily $COUNT)


}
\keyword{datasets}
