% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_green_fritillary_weekly.R
\docType{data}
\name{dark_green_fritillary_weekly}
\alias{dark_green_fritillary_weekly}
\title{Dark Green Fritillary Weekly}
\format{
A dataframe with 5 columns and 82420rows. The columns are WEEKNO (week number), SITENO (site number), YEAR (year), SPECIES (species identify), and COUNT (count).
}
\usage{
data(dark_green_fritillary_weekly)
}
\description{
Weekly site, year, and count data for dark green fritillary
}
\examples{

library(ButterflyLifespan)

data(dark_green_fritillary_weekly)
head(dark_green_fritillary_weekly)
plot(dark_green_fritillary_weekly$COUNT)


}
\keyword{datasets}
