% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marbled_white_daily.R
\docType{data}
\name{marbled_white_daily}
\alias{marbled_white_daily}
\title{Marbled White}
\format{
A dataframe with 5 columns and 1048575 rows. The columns are DAYNO (day number), WEEKNO (week number), SITENO (site number), YEAR (year), SPECIES (species identify), COUNT (count), and DATE (date).
}
\usage{
data(marbled_white_daily)
}
\description{
Daily site, year, and count data for marbled white
}
\examples{

library(ButterflyLifespan)

data(marbled_white_daily)
head(marbled_white_daily)
plot(marbled_white_daily$COUNT)


}
\keyword{datasets}
