% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marbled_white_weekly.R
\docType{data}
\name{marbled_white_weekly}
\alias{marbled_white_weekly}
\title{Marbled White Weekly}
\format{
A dataframe with 5 columns and 206778 rows. The columns are WEEKNO (week number), SITENO (site number), YEAR (year), SPECIES (species identify), and COUNT (count).
}
\usage{
data(marbled_white_weekly)
}
\description{
Weekly site, year, and count data for marbled white
}
\examples{

library(ButterflyLifespan)

data(marbled_white_weekly)
head(marbled_white_weekly)
plot(marbled_white_weekly$COUNT)


}
\keyword{datasets}
