% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orange_tip_daily.R
\docType{data}
\name{orange_tip_daily}
\alias{orange_tip_daily}
\title{Orange Tip Daily}
\format{
A dataframe with 5 columns and 1048575 rows. The columns are DAYNO (day number), WEEKNO (week number), SITENO (site number), YEAR (year), SPECIES (species identify), COUNT (count), and DATE (date).
}
\usage{
data(orange_tip_daily)
}
\description{
Daily site, year, and count data for orange tip
}
\examples{

library(ButterflyLifespan)

data(orange_tip_daily)
head(orange_tip_daily)
plot(orange_tip_daily$COUNT)


}
\keyword{datasets}
