% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orange_tip_weekly.R
\docType{data}
\name{orange_tip_weekly}
\alias{orange_tip_weekly}
\title{Orange Tip Weekly}
\format{
A dataframe with 5 columns and 346164 rows. The columns are WEEKNO (week number), SITENO (site number), YEAR (year), SPECIES (species identify), and COUNT (count).
}
\usage{
data(orange_tip_weekly)
}
\description{
Weekly site, year, and count data for orange tip
}
\examples{

library(ButterflyLifespan)

data(orange_tip_weekly)
head(orange_tip_weekly)
plot(orange_tip_weekly$COUNT)


}
\keyword{datasets}
