% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{TODO - Brief description}
\usage{
simulation(
  setup_level,
  env_cov.type,
  a.type,
  dist.type,
  B,
  cov_sim_type,
  n.iter,
  dpw,
  meth,
  convert_weekly,
  vary_phi,
  NA_type,
  nS,
  year
)
}
\arguments{
\item{setup_level}{character value "daily" or "weekly" dependent on how the data is to be sampled.}

\item{env_cov.type}{how to add environmental covariate data, either "lin", "quad" or "none"}

\item{a.type}{"N", "SO" or "S"}

\item{dist.type}{"P", "NB" or "ZIP"}

\item{B}{number of broods}

\item{cov_sim_type}{how covariate data is simulated - "uniform" or "seasonal"}

\item{n.iter}{number of iterations}

\item{dpw}{number of days per week to simulate}

\item{meth}{optimisation method - "Nelder-Mead","SANN" or "L-BFGS-B"}

\item{convert_weekly}{convert from the daily to the weekly format -  "yes" or "no"}

\item{vary_phi}{vary phi over the season -  "yes" or "no"}

\item{NA_type}{missing values - "match_NA", "prop_NA" or "none"}

\item{nS}{number of sites, if matching raw data then use the length of the output}
}
\value{
TODO  describe the ouput structure here
}
\description{
TODO - Longer description here (maybe one or two sentences)
}
\examples{

\donttest{
library(ButterflyLS)

# setup simulation parameters
year <- "2018"
setup_level <- "daily"
env_cov.type <- "none"
a.type <- "SO" 
dist.type <- "P" 
B <- 1   ## single brood
cov_sim_type <- "uniform"
n.iter <-250
dpw <- 1
meth <- "Nelder-Mead"
convert_weekly <- "no"
vary_phi <- "no"
NA_type <- "prop_NA"
nS <- 200

# run simulation
results <- simulation(setup_level,
		         env_cov.type,
		         a.type,
		         dist.type, 
		         B,
		         cov_sim_type,
		         n.iter,
		         dpw,
		         meth,
		         convert_weekly,
		         vary_phi,
		         NA_type,
		         nS,
		        year)
}

}
