% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Analysis code for fitting a GAI stopover model to a single year of data.}
\usage{
analysis(data, year, setup_level)
}
\arguments{
\item{data}{dataframe containing the data to analysis}

\item{year}{character value corresponding to the the year to analyse from the data}

\item{setup_level}{character value "daily" or "weekly" dependent on how the data was sampled.}
}
\value{
A list object containing three lists. The first containing the output used for deriving lifespan estimates and the second containing each output from each random start. The third states the likelihood for each random start.
}
\description{
Code used for fitting a single year GAI stopover model. The phi parameter is used to derive lifespan estimates. Data can be added either in a "daily" or "weekly" format.
}
\examples{

\donttest{
library(ButterflyLS)

# lazy load data
data(dark_green_fritillary_weekly)
# year to analyse
year <- "2018"
# weekly sampling
setup_level <- "weekly"

# run analysis
results <- analysis(dark_green_fritillary_weekly,year,setup_level)
}

}
