% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_multiyear.R
\name{analysis_multiyear}
\alias{analysis_multiyear}
\title{Analysis code for fitting a multiyear GAI stopover model to data spanning multiple years.}
\usage{
analysis_multiyear(data, setup_level, phi_type)
}
\arguments{
\item{data}{dataframe containing the data to analysis}

\item{setup_level}{character value "daily" or "weekly" dependent on how the data was sampled.}

\item{phi_type}{character value "doubleslope","slope", "variable" or "fixed"}
}
\value{
A list object containing three lists. The first containing the output used for deriving lifespan estimates and the second containing each output from each random start. The third states the likelihood for each random start.
}
\description{
Code used for fitting a multiyear GAI stopover model to data spanning multiple years. The phi parameter is used to derive lifespan estimates and can be modelled to be "fixed" across all years, to be fully "variable" year-to-year or modelled to change gradually with a "slope" term. Data can be added either in a "daily" or "weekly" format.
}
\examples{

\donttest{
library(ButterflyLS)

# lazy load data
data(orange_tip_weekly)
# year to analyse
year <- "2018"
# weekly sampling
setup_level <- "weekly"
# phi type
phi_type = "slope"

# run analysis
results <- analysis_multiyear(orange_tip_weekly,setup_level,phi_type)
}

}
